Move rfc6962 hasher out
