+++ merkle/rfc6962/rfc6962.go
@@ -1,4 +1,4 @@
-// Copyright 2016 Google LLC. All Rights Reserved.
+// Copyright 2021 Google LLC. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,10 +17,10 @@ package rfc6962
 
 import (
 	"crypto"
-	_ "crypto/sha256" // SHA256 is the default algorithm.
 
 	"github.com/google/trillian"
 	"github.com/google/trillian/merkle/hashers/registry"
+	"github.com/google/trillian/merkle/rfc6962/hasher"
 )
 
 func init() {
@@ -28,59 +28,26 @@ func init() {
 }
 
 // Domain separation prefixes
+//
+// Deprecated: moved to hasher subpackage.
 const (
-	RFC6962LeafHashPrefix = 0
-	RFC6962NodeHashPrefix = 1
+	RFC6962LeafHashPrefix = hasher.RFC6962LeafHashPrefix
+	RFC6962NodeHashPrefix = hasher.RFC6962NodeHashPrefix
 )
 
 // DefaultHasher is a SHA256 based LogHasher.
-var DefaultHasher = New(crypto.SHA256)
+//
+// Deprecated: moved to hasher subpackage.
+var DefaultHasher = hasher.DefaultHasher
 
 // Hasher implements the RFC6962 tree hashing algorithm.
-type Hasher struct {
-	crypto.Hash
-}
+//
+// Deprecated: moved to hasher subpackage.
+type Hasher = hasher.Hasher
 
 // New creates a new Hashers.LogHasher on the passed in hash function.
+//
+// Deprecated: moved to hasher subpackage.
 func New(h crypto.Hash) *Hasher {
-	return &Hasher{Hash: h}
-}
-
-// EmptyRoot returns a special case for an empty tree.
-func (t *Hasher) EmptyRoot() []byte {
-	return t.New().Sum(nil)
-}
-
-// HashLeaf returns the Merkle tree leaf hash of the data passed in through leaf.
-// The data in leaf is prefixed by the LeafHashPrefix.
-func (t *Hasher) HashLeaf(leaf []byte) []byte {
-	h := t.New()
-	h.Write([]byte{RFC6962LeafHashPrefix})
-	h.Write(leaf)
-	return h.Sum(nil)
-}
-
-// hashChildrenOld returns the inner Merkle tree node hash of the two child nodes l and r.
-// The hashed structure is NodeHashPrefix||l||r.
-// TODO(al): Remove me.
-func (t *Hasher) hashChildrenOld(l, r []byte) []byte {
-	h := t.New()
-	h.Write([]byte{RFC6962NodeHashPrefix})
-	h.Write(l)
-	h.Write(r)
-	return h.Sum(nil)
-}
-
-// HashChildren returns the inner Merkle tree node hash of the two child nodes l and r.
-// The hashed structure is NodeHashPrefix||l||r.
-func (t *Hasher) HashChildren(l, r []byte) []byte {
-	h := t.New()
-	b := append(append(append(
-		make([]byte, 0, 1+len(l)+len(r)),
-		RFC6962NodeHashPrefix),
-		l...),
-		r...)
-
-	h.Write(b)
-	return h.Sum(nil)
+	return hasher.New(h)
 }
